Refactor library into modular structure and add roadmap
- Reorganized monolithic lib.rs into domain-specific modules
- Enhanced documentation with UTF-8 mathematical notation
- Expanded mathematical traits with proper definitions
- Added features section in Cargo.toml for future extensions
- Created ROADMAP.md with focus on abstract algebraic structures
- Added support for advanced linear algebra and tensor concepts
