+++ Cargo.toml
@@ -8,12 +8,16 @@ license = "MIT"
 readme = "README.md"
 repository = "https://github.com/warlock-labs/noether"
 name = "noether"
-version = "0.2.0"
+version = "0.3.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 num-traits = "0.2.19"
+
+[features]
+default = []
+# Placeholder for future unstable features
 
 [lib]
