+++ ROADMAP.md
@@ -1,0 +1,89 @@
+# Roadmap
+
+Nœther is a library dedicated to abstract algebraic structures and their relationships. It focuses exclusively on defining traits and their mathematical properties, not on concrete implementations. This roadmap outlines the future development of Nœther as a framework for abstract algebra in Rust.
+
+## Abstract Algebraic Structures
+
+### Core Algebraic Traits
+- Complete marker traits for all fundamental algebraic properties
+- Enhance existing trait definitions with more complete axiomatization
+- Add trait methods that capture essential algebraic operations
+
+### Expanded Algebraic Hierarchy
+- Define traits for advanced algebraic structures:
+  - Lattices and Boolean Algebras
+  - Ordered Algebraic Structures
+  - Semirings and Dioids
+  - Bialgebras and Hopf Algebras
+
+### Category Theory Foundations
+- Add traits for categorical concepts:
+  - Categories, Functors, and Natural Transformations
+  - Monads and Comonads
+  - Limits and Colimits
+  - Adjunctions
+
+### Advanced Algebraic Systems
+- Extend to higher algebraic structures:
+  - Clifford Algebras and Geometric Algebra
+  - Lie Algebras and Lie Groups
+  - Jordan Algebras
+  - Quantum Groups and Quantum Algebras
+
+## Advanced Mathematical Domains
+
+### Abstract Linear Algebra
+- Enhance vector space abstraction with:
+  - Dual spaces and dual basis concepts
+  - Quotient spaces and subspace traits
+  - Universal properties of linear constructions
+  - Multilinear maps and tensor algebra
+
+### Algebraic Topology Concepts
+- Define traits for:
+  - Homology and Cohomology theories
+  - Chain complexes
+  - Simplicial and CW complexes
+  - Spectral sequences
+
+### Differential Structures
+- Abstract definitions for:
+  - Differential manifolds
+  - Tangent and cotangent bundles
+  - Differential forms
+  - Connections and curvature
+
+## Library Enhancement
+
+### Documentation and Theory
+- Expand mathematical documentation with formal definitions and theorems
+- Add more examples of how abstract structures relate to concrete mathematics
+- Develop visual guides for the trait hierarchy
+- Include references to mathematical literature for each concept
+
+### Type System Improvements
+- Leverage advanced Rust features for more expressive trait definitions:
+  - Const generics for dimension and rank specifications
+  - Associated type constructors for higher-kinded abstractions
+  - Type-level programming for compile-time proofs of properties
+
+### Verification and Testing
+- Develop property-based tests that verify mathematical laws:
+  - Tests for associativity, commutativity, distributivity, etc.
+  - Verify categorical properties and universal constructions
+  - Ensure trait implementations follow mathematical axioms
+
+### Ecosystem Integration
+- Define conversion traits as interfaces to concrete implementations
+- Provide trait adapters for existing Rust math libraries
+- Create guidelines for implementing Nœther traits in domain-specific libraries
+
+## Implementation Priorities
+
+The immediate focus is on:
+1. Completing and refining core algebraic traits
+2. Enhancing documentation with formal mathematical definitions
+3. Developing property-based testing for mathematical laws
+4. Expanding to advanced algebraic structures
+
+Nœther will remain true to its purpose as a library of abstract algebraic structures, providing the mathematical foundation upon which concrete implementations can be built in other libraries.
