+++ src/fields.rs
@@ -1,0 +1,121 @@
+//! Field theory structures.
+//!
+//! This module defines traits for algebraic structures related to fields,
+//! including Euclidean domains, fields, and field extensions.
+
+use crate::groups::MultiplicativeAbelianGroup;
+use crate::rings::PrincipalIdealDomain;
+use crate::spaces::VectorSpace;
+use num_traits::{Euclid, One, Zero};
+
+/// Represents a Euclidean Domain, an integral domain with a Euclidean function.
+///
+/// # Mathematical Definition
+/// A Euclidean Domain (R, +, ·, φ) is a principal ideal domain equipped with a
+/// Euclidean function φ: R\{0} → ℕ₀ that satisfies certain properties.
+///
+/// # Formal Definition
+/// Let (R, +, ·) be an integral domain and φ: R\{0} → ℕ₀ a function. R is a Euclidean domain if:
+/// 1. ∀a, b ∈ R, b ≠ 0, ∃!q, r ∈ R : a = bq + r ∧ (r = 0 ∨ φ(r) < φ(b)) (Division with Remainder)
+/// 2. ∀a, b ∈ R\{0} : φ(a) ≤ φ(ab) (Multiplicative Property)
+pub trait EuclideanDomain: PrincipalIdealDomain + Euclid {}
+
+/// Represents a Field, a commutative ring where every non-zero element has a multiplicative inverse.
+///
+/// # Mathematical Definition
+/// A field (F, +, ·) is a commutative ring where:
+/// - Every non-zero element has a multiplicative inverse
+///
+/// # Formal Definition
+/// Let (F, +, ·) be a field. Then:
+/// 1. (F, +, ·) is a commutative ring
+/// 2. ∀ a ∈ F, a ≠ 0, ∃ a⁻¹ ∈ F, a · a⁻¹ = a⁻¹ · a = 1 (multiplicative inverse)
+pub trait Field: EuclideanDomain + MultiplicativeAbelianGroup {}
+
+/// Represents a Finite Field, a field with a finite number of elements.
+///
+/// # Mathematical Definition
+/// A finite field is a field with a finite number of elements.
+///
+/// # Properties
+/// - The number of elements is always a prime power p^n
+pub trait FiniteField: Field {
+    ///Allows for an arbitrary size representation without specificing type
+    type ScalarType: Clone + PartialOrd + Zero + One;
+
+    /// Returns the characteristic of the field.
+    fn characteristic() -> u64;
+
+    /// Returns the number of elements in the field.
+    fn order() -> u64;
+}
+
+/// Represents an Ordered Field, a field with a total order compatible with its operations.
+///
+/// # Mathematical Definition
+/// An ordered field is a field equipped with a total order ≤ where:
+/// - If a ≤ b then a + c ≤ b + c for all c
+/// - If 0 ≤ a and 0 ≤ b then 0 ≤ a · b
+pub trait OrderedField: Field + PartialOrd {}
+
+/// Represents a Real Field, a complete ordered field.
+///
+/// # Mathematical Definition
+/// A real field is an ordered field that is Dedekind-complete:
+/// - Every non-empty subset with an upper bound has a least upper bound
+pub trait RealField: OrderedField {}
+
+/// Represents a Field Extension.
+///
+/// # Mathematical Definition
+/// A field extension L/K is a field L that contains K as a subfield.
+///
+/// # Properties
+/// - L is a field
+/// - K is a subfield of L
+/// - L is a vector space over K
+pub trait FieldExtension: Field + VectorSpace<Self::BaseField> {
+    /// The base field of this extension.
+    type BaseField: Field;
+
+    /// Returns the degree of the field extension.
+    fn degree() -> usize;
+
+    /// Computes the trace of an element.
+    fn trace(&self) -> Self::BaseField;
+
+    /// Computes the norm of an element.
+    fn norm(&self) -> Self::BaseField;
+}
+
+/// Represents a Tower of Field Extensions.
+///
+/// # Mathematical Definition
+/// A tower of field extensions is a sequence of fields K = F₀ ⊂ F₁ ⊂ ... ⊂ Fₙ = L
+/// where each Fᵢ₊₁/Fᵢ is a field extension.
+///
+/// # Properties
+/// - Each level is a field extension of the previous level
+/// - The composition of the extensions forms the overall extension L/K
+/// - The degree of L/K is the product of the degrees of each extension in the tower
+pub trait FieldExtensionTower: FieldExtension {
+    /// Returns the number of extensions in the tower.
+    fn tower_height() -> usize;
+
+    /// Returns the degree of the i-th extension in the tower.
+    fn extension_degree(i: usize) -> usize;
+}
+
+// Blanket implementations
+impl<T: PrincipalIdealDomain + Euclid> EuclideanDomain for T {}
+impl<T: EuclideanDomain + MultiplicativeAbelianGroup> Field for T {}
+impl<T: Field + PartialOrd> OrderedField for T {}
+
+// RealField
+// Note: This cannot be implemented as a blanket impl because it requires knowledge about completeness
+
+// FieldExtension
+// Note: This cannot be implemented as a blanket impl because it requires specific knowledge about the base field and extension structure
+
+// FieldExtensionTower
+// Note: This cannot be implemented as a blanket impl because it requires specific knowledge about the tower structure
