+++ src/sets.rs
@@ -1,0 +1,32 @@
+//! Set theory foundations.
+//!
+//! This module provides the foundational Set trait which forms the basis
+//! of all algebraic structures in Noether.
+
+/// Represents a mathematical set as defined in Zermelo-Fraenkel set theory with Choice (ZFC).
+///
+/// # Formal Notation
+/// - ∅: empty set
+/// - ∈: element of
+/// - ⊆: subset of
+/// - ∪: union
+/// - ∩: intersection
+/// - \: set difference
+/// - Δ: symmetric difference
+/// - |A|: cardinality of set A
+///
+/// # Axioms of ZFC
+/// 1. Extensionality: ∀A∀B(∀x(x ∈ A ↔ x ∈ B) → A = B)
+/// 2. Empty Set: ∃A∀x(x ∉ A)
+/// 3. Pairing: ∀a∀b∃A∀x(x ∈ A ↔ x = a ∨ x = b)
+/// 4. Union: ∀F∃A∀x(x ∈ A ↔ ∃B(x ∈ B ∧ B ∈ F))
+/// 5. Power Set: ∀A∃P∀x(x ∈ P ↔ x ⊆ A)
+/// 6. Infinity: ∃A(∅ ∈ A ∧ ∀x(x ∈ A → x ∪ {x} ∈ A))
+/// 7. Separation: ∀A∃B∀x(x ∈ B ↔ x ∈ A ∧ φ(x)) for any formula φ
+/// 8. Replacement: ∀A(∀x∀y∀z((x ∈ A ∧ φ(x,y) ∧ φ(x,z)) → y = z) → ∃B∀y(y ∈ B ↔ ∃x(x ∈ A ∧ φ(x,y))))
+/// 9. Foundation: ∀A(A ≠ ∅ → ∃x(x ∈ A ∧ x ∩ A = ∅))
+/// 10. Choice: ∀A(∅ ∉ A → ∃f:A → ∪A ∀B∈A(f(B) ∈ B))
+pub trait Set: Sized + Clone + PartialEq {}
+
+// Blanket implementation for any type that satisfies the trait bounds
+impl<T: Clone + PartialEq> Set for T {}
